feat: implement tasks from "modules" module
