+++ src/modules/esm.mjs
@@ -1,0 +1,42 @@
+import path, { dirname } from "path";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import { fileURLToPath } from "url";
+
+import "./files/c.js";
+
+const random = Math.random();
+
+let unknownObject;
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+if (random > 0.5) {
+    unknownObject = await import(path.join(__dirname, "files", "a.json"), { with: { type: "json" } });
+} else {
+    unknownObject = await import(path.join(__dirname, "files", "b.json"), { with: { type: "json" } });
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${import.meta.url}`);
+console.log(`Path to current directory is ${path.dirname(import.meta.url)}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
+
